+++ R/airsis_ESAMQualityControl.R
@@ -1,7 +1,7 @@
 #' @keywords AIRSIS
 #' @export
 #' @title Apply Quality Control to Raw AIRSIS E-Sampler Dataframe
-#' @param df single site dataframe created by \code{airsis_downloadData()}
+#' @param tbl single site tibble created by \code{airsis_downloadData()}
 #' @param valid_Longitude range of valid Longitude values
 #' @param valid_Latitude range of valid Latitude values
 #' @param remove_Lon_zero flag to remove rows where Longitude == 0
@@ -23,10 +23,10 @@
 #' 
 #' A \code{POSIXct datetime} column (UTC) is also added based on \code{TimeStamp}.
 #' 
-#' @return Cleaned up dataframe of AIRSIS monitor data.
+#' @return Cleaned up tibble of AIRSIS monitor data.
 #' @seealso \code{\link{airsis_qualityControl}}
 
-airsis_ESAMQualityControl <- function(df,
+airsis_ESAMQualityControl <- function(tbl,
                                       valid_Longitude=c(-180,180),
                                       valid_Latitude=c(-90,90),
                                       remove_Lon_zero = TRUE,
@@ -39,7 +39,7 @@ airsis_ESAMQualityControl <- function(df,
   
   # TODO:  What about Alarm?
   
-  #   > names(df)
+  #   > names(tbl)
   #    [1] "MasterTable_ID"        "Alias"                 "Latitude"              "Longitude"            
   #    [5] "Conc.mg.m3."           "Flow.l.m."             "AT.C."                 "BP.PA."               
   #    [9] "RHx..."                "RHi..."                "WS.M.S."               "WD.Deg."              
@@ -47,7 +47,7 @@ airsis_ESAMQualityControl <- function(df,
   #   [17] "System.Volts"          "Data.1"                "Data.2"                "TimeStamp"            
   #   [21] "PDate"                 "monitorName"           "monitorType"          
   
-  monitorName <- df$monitorName[1]
+  monitorName <- tbl$monitorName[1]
   
   # ----- Missing Values ------------------------------------------------------
   
@@ -59,21 +59,21 @@ airsis_ESAMQualityControl <- function(df,
     # verb for logging messages
     verb <- "Flagging"
     
-    df$rowID <- as.integer(rownames(df))
+    tbl$rowID <- as.integer(rownames(tbl))
     
-    # duplicate df and add columns for flags
-    dfFlagged <- df
-    dfFlagged$QCFlag_anyBad <- FALSE
-    dfFlagged$QCFlag_reasonCode <- NA
-    dfFlagged$QCFlag_badLon <- FALSE
-    dfFlagged$QCFlag_badLat <- FALSE
-    dfFlagged$QCFlag_badType <- FALSE # no type info for ESAMs
-    dfFlagged$QCFlag_badFlow <- FALSE
-    dfFlagged$QCFlag_badAT <- FALSE
-    dfFlagged$QCFlag_badRHi <- FALSE
-    dfFlagged$QCFlag_badConcHr <- FALSE
-    dfFlagged$QCFlag_badDateAndTime <- FALSE
-    dfFlagged$QCFlag_duplicateHr <- FALSE
+    # duplicate tbl and add columns for flags
+    tblFlagged <- tbl
+    tblFlagged$QCFlag_anyBad <- FALSE
+    tblFlagged$QCFlag_reasonCode <- NA
+    tblFlagged$QCFlag_badLon <- FALSE
+    tblFlagged$QCFlag_badLat <- FALSE
+    tblFlagged$QCFlag_badType <- FALSE # no type info for ESAMs
+    tblFlagged$QCFlag_badFlow <- FALSE
+    tblFlagged$QCFlag_badAT <- FALSE
+    tblFlagged$QCFlag_badRHi <- FALSE
+    tblFlagged$QCFlag_badConcHr <- FALSE
+    tblFlagged$QCFlag_badDateAndTime <- FALSE
+    tblFlagged$QCFlag_duplicateHr <- FALSE
   } else {
     # verb for logging messages
     verb <- "Discarding"
@@ -83,35 +83,35 @@ airsis_ESAMQualityControl <- function(df,
   
   # Latitude and longitude must be in range
   if ( remove_Lon_zero ) {
-    goodLonMask <- !is.na(df$Longitude) & (df$Longitude >= valid_Longitude[1]) & (df$Longitude <= valid_Longitude[2]) & (df$Longitude != 0)
+    goodLonMask <- !is.na(tbl$Longitude) & (tbl$Longitude >= valid_Longitude[1]) & (tbl$Longitude <= valid_Longitude[2]) & (tbl$Longitude != 0)
   } else {
-    goodLonMask <- !is.na(df$Longitude) & (df$Longitude >= valid_Longitude[1]) & (df$Longitude <= valid_Longitude[2])
+    goodLonMask <- !is.na(tbl$Longitude) & (tbl$Longitude >= valid_Longitude[1]) & (tbl$Longitude <= valid_Longitude[2])
   }
   
   if ( remove_Lat_zero ) {
-    goodLatMask <- !is.na(df$Latitude) & (df$Latitude >= valid_Latitude[1]) & (df$Latitude <= valid_Latitude[2]) & (df$Latitude != 0)
+    goodLatMask <- !is.na(tbl$Latitude) & (tbl$Latitude >= valid_Latitude[1]) & (tbl$Latitude <= valid_Latitude[2]) & (tbl$Latitude != 0)
   } else {    
-    goodLatMask <- !is.na(df$Latitude) & (df$Latitude >= valid_Latitude[1]) & (df$Latitude <= valid_Latitude[2])
+    goodLatMask <- !is.na(tbl$Latitude) & (tbl$Latitude >= valid_Latitude[1]) & (tbl$Latitude <= valid_Latitude[2])
   }
   
   badRows <- !(goodLonMask & goodLatMask)
   badRowCount <- sum(badRows)
   if ( badRowCount > 0 ) {
     logger.info(paste(verb,"%s rows with invalid location information"), badRowCount)
-    badLocations <- paste('(',df$Longitude[badRows],',',df$Latitude[badRows],')',sep='')
+    badLocations <- paste('(',tbl$Longitude[badRows],',',tbl$Latitude[badRows],')',sep='')
     logger.debug("Bad locations: %s", paste0(badLocations, collapse=", "))
     if ( flagAndKeep ) {
       # apply flags
-      dfFlagged$QCFlag_badLon[df$rowID[!goodLonMask]] <- TRUE
-      dfFlagged$QCFlag_badLat[df$rowID[!goodLatMask]] <- TRUE
-      dfFlagged$QCFlag_anyBad <- dfFlagged$QCFlag_anyBad | dfFlagged$QCFlag_badLon | dfFlagged$QCFlag_badLat
+      tblFlagged$QCFlag_badLon[tbl$rowID[!goodLonMask]] <- TRUE
+      tblFlagged$QCFlag_badLat[tbl$rowID[!goodLatMask]] <- TRUE
+      tblFlagged$QCFlag_anyBad <- tblFlagged$QCFlag_anyBad | tblFlagged$QCFlag_badLon | tblFlagged$QCFlag_badLat
       # apply reason codes
-      dfFlagged$QCFlag_reasonCode[df$rowID[!goodLonMask]] <- paste(dfFlagged$QCFlag_reasonCode[df$rowID[!goodLonMask]],"badLon")
-      dfFlagged$QCFlag_reasonCode[df$rowID[!goodLatMask]] <- paste(dfFlagged$QCFlag_reasonCode[df$rowID[!goodLatMask]],"badLat")
+      tblFlagged$QCFlag_reasonCode[tbl$rowID[!goodLonMask]] <- paste(tblFlagged$QCFlag_reasonCode[tbl$rowID[!goodLonMask]],"badLon")
+      tblFlagged$QCFlag_reasonCode[tbl$rowID[!goodLatMask]] <- paste(tblFlagged$QCFlag_reasonCode[tbl$rowID[!goodLatMask]],"badLat")
     }
   }
   
-  df <- df[goodLonMask & goodLatMask,]
+  tbl <- tbl[goodLonMask & goodLatMask,]
   
   # ----- Time ----------------------------------------------------------------
   
@@ -123,13 +123,13 @@ airsis_ESAMQualityControl <- function(df,
   # NOTE: actually a record for 4:00pm (which is really representative of data during the 3:00 hour -- see NOTE below).
   
   # Add a POSIXct datetime
-  df$datetime <- lubridate::floor_date(lubridate::mdy_hms(df$TimeStamp), unit="hour") - lubridate::dhours(1)
+  tbl$datetime <- lubridate::floor_date(lubridate::mdy_hms(tbl$TimeStamp), unit="hour") - lubridate::dhours(1)
   if ( flagAndKeep ) {
-    # TODO: Unable to get datetime moved from df to dfFlagged without timezone and/or display getting messed up.
+    # TODO: Unable to get datetime moved from tbl to tblFlagged without timezone and/or display getting messed up.
     # For now just duplicating the calculation, then assigning row values to NA after the fact for rows that were
-    # removed from df prior to calculating datetime above. Clean up later if possible.
-    dfFlagged$datetime <- lubridate::floor_date(lubridate::mdy_hms(dfFlagged$TimeStamp), unit="hour") - lubridate::dhours(1)
-    dfFlagged$datetime[ which(!(dfFlagged$rowID %in% df$rowID)) ] <- NA
+    # removed from tbl prior to calculating datetime above. Clean up later if possible.
+    tblFlagged$datetime <- lubridate::floor_date(lubridate::mdy_hms(tblFlagged$TimeStamp), unit="hour") - lubridate::dhours(1)
+    tblFlagged$datetime[ which(!(tblFlagged$rowID %in% tbl$rowID)) ] <- NA
   }
   
   # NOTE: The time above truncates the timestamp to the top of an hour, and then subtracts one hour,
@@ -157,22 +157,22 @@ airsis_ESAMQualityControl <- function(df,
   ###                                         ifelse(concQA == 'Negative', 0,
   ###                                         ifelse(concQA == 'OK', ConcHr, NA)))
   
-  goodFlow <- !is.na(df$Flow.l.m.) & df$Flow.l.m. >= valid_Flow[1] & df$Flow.l.m. <= valid_Flow[2]
-  goodAT <- !is.na(df$AT.C.) & df$AT.C. >= valid_AT[1] & df$AT.C. <= valid_AT[2]
-  goodRHi <- !is.na(df$RHi...) & df$RHi... >= valid_RHi[1] & df$RHi... <= valid_RHi[2]
-  goodConcHr <- !is.na(df$Conc.mg.m3.) & df$Conc.mg.m3. >= valid_Conc[1] & df$Conc.mg.m3. <= valid_Conc[2]
-  gooddatetime <- !is.na(df$datetime) & df$datetime < lubridate::now("UTC") # saw a future date once
+  goodFlow <- !is.na(tbl$Flow.l.m.) & tbl$Flow.l.m. >= valid_Flow[1] & tbl$Flow.l.m. <= valid_Flow[2]
+  goodAT <- !is.na(tbl$AT.C.) & tbl$AT.C. >= valid_AT[1] & tbl$AT.C. <= valid_AT[2]
+  goodRHi <- !is.na(tbl$RHi...) & tbl$RHi... >= valid_RHi[1] & tbl$RHi... <= valid_RHi[2]
+  goodConcHr <- !is.na(tbl$Conc.mg.m3.) & tbl$Conc.mg.m3. >= valid_Conc[1] & tbl$Conc.mg.m3. <= valid_Conc[2]
+  gooddatetime <- !is.na(tbl$datetime) & tbl$datetime < lubridate::now("UTC") # saw a future date once
   
   logger.debug("Flow has %s missing or out of range values", sum(!goodFlow))
-  if (sum(!goodFlow) > 0) logger.debug("Bad Flow values:  %s", paste0(sort(unique(df$Flow.l.m.[!goodFlow]),na.last=TRUE), collapse=", "))
+  if (sum(!goodFlow) > 0) logger.debug("Bad Flow values:  %s", paste0(sort(unique(tbl$Flow.l.m.[!goodFlow]),na.last=TRUE), collapse=", "))
   logger.debug("AT has %s missing or out of range values", sum(!goodAT))
-  if (sum(!goodAT) > 0) logger.debug("Bad AT values:  %s", paste0(sort(unique(df$AT.C.[!goodAT]),na.last=TRUE), collapse=", "))
+  if (sum(!goodAT) > 0) logger.debug("Bad AT values:  %s", paste0(sort(unique(tbl$AT.C.[!goodAT]),na.last=TRUE), collapse=", "))
   logger.debug("RHi has %s missing or out of range values", sum(!goodRHi))
-  if (sum(!goodRHi) > 0) logger.debug("Bad RHi values:  %s", paste0(sort(unique(df$RHi...[!goodRHi]),na.last=TRUE), collapse=", "))
+  if (sum(!goodRHi) > 0) logger.debug("Bad RHi values:  %s", paste0(sort(unique(tbl$RHi...[!goodRHi]),na.last=TRUE), collapse=", "))
   logger.debug("Conc has %s missing or out of range values", sum(!goodConcHr))
-  if (sum(!goodConcHr) > 0) logger.debug("Bad Conc values:  %s", paste0(sort(unique(df$Conc.mg.m3.[!goodConcHr]),na.last=TRUE), collapse=", "))
+  if (sum(!goodConcHr) > 0) logger.debug("Bad Conc values:  %s", paste0(sort(unique(tbl$Conc.mg.m3.[!goodConcHr]),na.last=TRUE), collapse=", "))
   logger.debug("datetime has %s missing or out of range values", sum(!gooddatetime))
-  if (sum(!gooddatetime) > 0) logger.debug("Bad datetime values:  %s", paste0(sort(unique(df$datetime[!gooddatetime]),na.last=TRUE), collapse=", "))
+  if (sum(!gooddatetime) > 0) logger.debug("Bad datetime values:  %s", paste0(sort(unique(tbl$datetime[!gooddatetime]),na.last=TRUE), collapse=", "))
   
   goodMask <- goodFlow & goodAT & goodRHi & goodConcHr & gooddatetime
   badQCCount <- sum(!goodMask)
@@ -181,23 +181,23 @@ airsis_ESAMQualityControl <- function(df,
     logger.info(paste(verb,"%s rows because of QC logic"), badQCCount)
     if ( flagAndKeep ) {
       # apply flags
-      dfFlagged$QCFlag_badFlow[df$rowID[!goodFlow]] <- TRUE
-      dfFlagged$QCFlag_badAT[df$rowID[!goodAT]] <- TRUE
-      dfFlagged$QCFlag_badRHi[df$rowID[!goodRHi]] <- TRUE
-      dfFlagged$QCFlag_badConcHr[df$rowID[!goodConcHr]] <- TRUE
-      dfFlagged$QCFlag_badDateAndTime[df$rowID[!gooddatetime]] <- TRUE
-      dfFlagged$QCFlag_anyBad <- (dfFlagged$QCFlag_anyBad | dfFlagged$QCFlag_badFlow | dfFlagged$QCFlag_badAT | 
-                                    dfFlagged$QCFlag_badRHi | dfFlagged$QCFlag_badConcHr | dfFlagged$QCFlag_badDateAndTime)
+      tblFlagged$QCFlag_badFlow[tbl$rowID[!goodFlow]] <- TRUE
+      tblFlagged$QCFlag_badAT[tbl$rowID[!goodAT]] <- TRUE
+      tblFlagged$QCFlag_badRHi[tbl$rowID[!goodRHi]] <- TRUE
+      tblFlagged$QCFlag_badConcHr[tbl$rowID[!goodConcHr]] <- TRUE
+      tblFlagged$QCFlag_badDateAndTime[tbl$rowID[!gooddatetime]] <- TRUE
+      tblFlagged$QCFlag_anyBad <- (tblFlagged$QCFlag_anyBad | tblFlagged$QCFlag_badFlow | tblFlagged$QCFlag_badAT | 
+                                    tblFlagged$QCFlag_badRHi | tblFlagged$QCFlag_badConcHr | tblFlagged$QCFlag_badDateAndTime)
       # apply reason codes
-      dfFlagged$QCFlag_reasonCode[df$rowID[!goodFlow]] <- paste(dfFlagged$QCFlag_reasonCode[df$rowID[!goodFlow]],"badFlow")
-      dfFlagged$QCFlag_reasonCode[df$rowID[!goodAT]] <- paste(dfFlagged$QCFlag_reasonCode[df$rowID[!goodAT]],"badAT")
-      dfFlagged$QCFlag_reasonCode[df$rowID[!goodRHi]] <- paste(dfFlagged$QCFlag_reasonCode[df$rowID[!goodRHi]],"badRHi")
-      dfFlagged$QCFlag_reasonCode[df$rowID[!goodConcHr]] <- paste(dfFlagged$QCFlag_reasonCode[df$rowID[!goodConcHr]],"badConcHr")
-      dfFlagged$QCFlag_reasonCode[df$rowID[!gooddatetime]] <- paste(dfFlagged$QCFlag_reasonCode[df$rowID[!gooddatetime]],"badDateAndTime")
+      tblFlagged$QCFlag_reasonCode[tbl$rowID[!goodFlow]] <- paste(tblFlagged$QCFlag_reasonCode[tbl$rowID[!goodFlow]],"badFlow")
+      tblFlagged$QCFlag_reasonCode[tbl$rowID[!goodAT]] <- paste(tblFlagged$QCFlag_reasonCode[tbl$rowID[!goodAT]],"badAT")
+      tblFlagged$QCFlag_reasonCode[tbl$rowID[!goodRHi]] <- paste(tblFlagged$QCFlag_reasonCode[tbl$rowID[!goodRHi]],"badRHi")
+      tblFlagged$QCFlag_reasonCode[tbl$rowID[!goodConcHr]] <- paste(tblFlagged$QCFlag_reasonCode[tbl$rowID[!goodConcHr]],"badConcHr")
+      tblFlagged$QCFlag_reasonCode[tbl$rowID[!gooddatetime]] <- paste(tblFlagged$QCFlag_reasonCode[tbl$rowID[!gooddatetime]],"badDateAndTime")
     }
   }
   
-  df <- df[goodMask,]
+  tbl <- tbl[goodMask,]
   
   # ----- Duplicate Hours -----------------------------------------------------
     
@@ -206,43 +206,43 @@ airsis_ESAMQualityControl <- function(df,
   # NOTE: want to build in functionality to ensure that the latest is picked if more than one exists
   # NOTE: (for example, if the data is not in order by timestamp for whatever reason)
    
-  dupHrMask <- duplicated(df$datetime,fromLast = TRUE)
+  dupHrMask <- duplicated(tbl$datetime,fromLast = TRUE)
   dupHrCount <- sum(dupHrMask)
   uniqueHrMask <- !dupHrMask
   
   if ( dupHrCount > 0 ) {
     logger.info(paste(verb,"%s duplicate time entries"), dupHrCount)
-    logger.debug("Duplicate Hours (may be >1 per timestamp):  %s", paste0(sort(unique(df$TimeStamp[dupHrMask])), collapse=", "))
+    logger.debug("Duplicate Hours (may be >1 per timestamp):  %s", paste0(sort(unique(tbl$TimeStamp[dupHrMask])), collapse=", "))
     if ( flagAndKeep ) {
       # apply flags
-      dfFlagged$QCFlag_duplicateHr[df$rowID[dupHrMask]] <- TRUE
-      dfFlagged$QCFlag_anyBad <- dfFlagged$QCFlag_anyBad | dfFlagged$QCFlag_duplicateHr
+      tblFlagged$QCFlag_duplicateHr[tbl$rowID[dupHrMask]] <- TRUE
+      tblFlagged$QCFlag_anyBad <- tblFlagged$QCFlag_anyBad | tblFlagged$QCFlag_duplicateHr
       # apply reason code
-      dfFlagged$QCFlag_reasonCode[df$rowID[dupHrMask]] <- paste(dfFlagged$QCFlag_reasonCode[df$rowID[dupHrMask]],"duplicateHr")
+      tblFlagged$QCFlag_reasonCode[tbl$rowID[dupHrMask]] <- paste(tblFlagged$QCFlag_reasonCode[tbl$rowID[dupHrMask]],"duplicateHr")
     }
   }
 
-  df <- df[uniqueHrMask,]
+  tbl <- tbl[uniqueHrMask,]
   
   # ----- More QC -------------------------------------------------------------
   
   # NOTE:  Additional QC would go here
   
   if ( flagAndKeep ) {
-    logger.debug("Retaining %d rows of measurements; %d bad rows flagged", nrow(df), sum(dfFlagged$QCFlag_anyBad))    
+    logger.debug("Retaining %d rows of measurements; %d bad rows flagged", nrow(tbl), sum(tblFlagged$QCFlag_anyBad))    
   } else {
-    logger.debug("Retaining %d rows of validated measurements", nrow(df))
+    logger.debug("Retaining %d rows of validated measurements", nrow(tbl))
   }
   
   # ----- Final cleanup -------------------------------------------------------
   
   if ( flagAndKeep ) {
-    dfFlagged$QCFlag_reasonCode <- stringr::str_sub(dfFlagged$QCFlag_reasonCode, 3)
-    dfFlagged$QCFlag_reasonCode <- stringr::str_trim(dfFlagged$QCFlag_reasonCode)
-    df <- dfFlagged
-    df$rowID <- NULL
+    tblFlagged$QCFlag_reasonCode <- stringr::str_sub(tblFlagged$QCFlag_reasonCode, 3)
+    tblFlagged$QCFlag_reasonCode <- stringr::str_trim(tblFlagged$QCFlag_reasonCode)
+    tbl <- tblFlagged
+    tbl$rowID <- NULL
   }
   
-  return(df)
+  return(tbl)
   
 }